requirements.txt
# Core dependencies
streamlit>=1.29.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.10
openai>=1.12.0

# Vector store
chromadb>=0.4.22

# Data processing
pandas>=2.1.0

# Environment variables
python-dotenv>=1.0.0

# Text processing and embeddings
tiktoken>=0.5.0

# Additional utilities
pypdf>=3.17.0

# Web scraping for fallback content
beautifulsoup4>=4.12.0
requests>=2.31.0

# FastAPI backend
fastapi>=0.104.0
uvicorn>=0.24.0
